+++ src/build-data/version.txt
@@ -2,7 +2,7 @@
 release_major = 2
 release_minor = 11
 release_patch = 0
-release_so_abi_rev = 10
+release_so_abi_rev = 11
 
 # These are set by the distribution script
 release_vc_rev = None
