+++ src/lib/base/init.h
@@ -13,6 +13,8 @@
 
 namespace Botan {
 
+BOTAN_DEPRECATED_HEADER(init.h)
+
 /*
 * Previously botan had state whose lifetime had to be explicitly
 * managed by the application. As of 1.11.14 this is no longer the
