+++ src/lib/base/lookup.h
@@ -32,6 +32,8 @@
 
 namespace Botan {
 
+BOTAN_DEPRECATED_HEADER(lookup.h)
+
 /*
 * As of 1.11.26 this header is deprecated. Instead use the calls T::create and
 * T::providers (as demonstrated in the implementation below).
