+++ src/lib/block/threefish_512/threefish.h
@@ -8,6 +8,10 @@
 #ifndef BOTAN_THREEFISH_H_
 #define BOTAN_THREEFISH_H_
 
+// This header is deprecated and will be removed in a future major release
+
 #include <botan/threefish_512.h>
+
+BOTAN_DEPRECATED_HEADER(threefish.h)
 
 #endif
