+++ src/lib/filters/comp_filter.cpp
@@ -6,7 +6,7 @@
 * Botan is released under the Simplified BSD License (see license.txt)
 */
 
-#include <botan/comp_filter.h>
+#include <botan/filters.h>
 #include <botan/exceptn.h>
 
 #if defined(BOTAN_HAS_COMPRESSION)
