+++ src/lib/filters/threaded_fork.cpp
@@ -6,7 +6,7 @@
 * Botan is released under the Simplified BSD License (see license.txt)
 */
 
-#include <botan/basefilt.h>
+#include <botan/filters.h>
 
 #if defined(BOTAN_HAS_THREAD_UTILS)
