+++ src/lib/pubkey/sm2/sm2_enc.h
@@ -10,4 +10,6 @@
 
 #include <botan/sm2.h>
 
+BOTAN_DEPRECATED_HEADER(sm2_enc.h)
+
 #endif
